fix :alembic issue fixed

+++ backend/server/models/user.py
@@ -1,13 +1,14 @@
+from datetime import datetime
+
 import sqlalchemy as sa
-from .base import Base
-import datetime
+from db.database import Base
 
 
 class User(Base):
     __tablename__ = "users"
     id = sa.Column(sa.Integer(), primary_key=True)
     email = sa.Column(sa.String(255), unique=True)
-    phone_number = sa.Column(sa.Integer(10), unique=True)
+    phone_number = sa.Column(sa.Integer(), unique=True)
     password = sa.Column(sa.String(255))
     email = sa.Column(sa.String(255))
     created_on = sa.Column(sa.DateTime, default=datetime.utcnow)
